% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/composite.R
\docType{data}
\name{BioDB}
\alias{BioDB}
\title{Biomarker Thresholds Database}
\format{a \code{list}.}
\source{
Various publications
}
\description{
This data set lists the clinical high risk thresholds for a variety of biomarkers
}

