% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/composite.R
\name{distanceScores}
\alias{distanceScores}
\title{Calculate distance scores on data in preparation for composite scoring}
\usage{
distanceScores(d, g, thresholds, higherisbetter, winsorize = FALSE,
  better = TRUE, na.rm = TRUE, saveall = FALSE)
}
\arguments{
\item{d}{The data}

\item{g}{A grouping variable}

\item{thresholds}{Thresholds to use when calculating distances
(e.g., median, clinical thresholds, etc.).  If groups are used,
must be thresholds for each group (e.g., to allow separate thresholds for
females and males).}

\item{higherisbetter}{A logical vector for each biomarker whether higher scores
are better or not.}

\item{winsorize}{Whether to winsorize the data or not.  Defaults to \code{FALSE}.
If not \code{FALSE}, the percentile to winsorize at.  For example, .01 would be
the .01 and the 1 - .01 percentiles.}

\item{better}{Logical indicating whether \dQuote{better} values than the threshold are allowed.}

\item{na.rm}{A logical whether missing values should be ommitted}

\item{saveall}{A logical whether to save all intermediary datasets and graphs. Defaults to \code{FALSE}.}
}
\value{
A list of results.
}
\description{
Calculate distance scores on data in preparation for composite scoring
}
\examples{
# make me!
}

