% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/composite.R
\name{mahalanobisComposite}
\alias{mahalanobisComposite}
\title{Score Data Using the Mahalanobis Distance}
\usage{
mahalanobisComposite(object, ncomponents)
}
\arguments{
\item{object}{An object ready for use}

\item{ncomponents}{the number of components to use from the
principal component analysis}
}
\value{
A list of results.
}
\description{
Create Metabolic Syndrome Severity Score using the Mahalanobis Distance
}
\examples{
# make me!
}

