% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/composite.R
\name{prepareComposite}
\alias{prepareComposite}
\title{Prepare data to have a composite calculated}
\usage{
prepareComposite(object, covmat, standardize = TRUE)
}
\arguments{
\item{object}{An object ready for use}

\item{covmat}{The covariance matrix to use.  If missing,
austomatically calculated from the data.}

\item{standardize}{A logical value whether to standardize the data or not.}
}
\value{
A list of results.
}
\description{
Prepare data to have a composite calculated
}
\examples{
# make me!
}

