% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/composite.R
\name{scoreComposite}
\alias{scoreComposite}
\title{Score Data Using a simple sum}
\usage{
scoreComposite(object, transform = c("square", "abs", "none"))
}
\arguments{
\item{object}{An object ready for use}

\item{transformation}{A character string indicating the type of transformation to use.
One of \dQuote{square}, \dQuote{abs}, or \dQuote{none}, which either sums the raw data,
sums the squared data and then takes the square root, or sums the absolute values of the
data.}
}
\value{
A list of results.
}
\description{
Create Metabolic Syndrome Severity Score using summation
}
\examples{
# make me!
}

