% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/composite.R
\name{scoreFactor}
\alias{scoreFactor}
\title{Score Data Using a Factor Model}
\usage{
scoreFactor(object, type = c("onefactor", "secondorderfactor", "bifactor"),
  factors = "")
}
\arguments{
\item{object}{An object ready for use}

\item{type}{A character string indicating the type of factor model to use}

\item{factors}{A named list where names are the factor names and each
element is a character string of the indicator names.}
}
\value{
A list of results.
}
\description{
Create Metabolic Syndrome Severity Score using a Factor Model
}
\examples{
# make me!
}

