% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/composite.R
\name{winsorizor}
\alias{winsorizor}
\title{Winsorize at specified percentiles}
\usage{
winsorizor(d, percentile, na.rm = TRUE)
}
\arguments{
\item{d}{A vector, matrix, or data frame to be winsorized}

\item{percentile}{The percentile bounded by [0, 1] to winsorize data at}

\item{na.rm}{A logical whether to remove NAs.}
}
\value{
winsorized data.
}
\description{
Simple function winsorizes data at the specified percentile.
}
\examples{
#make me!
}

